LayoutTests/fast/worklets/worklet-constructor.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Tests that Worklet.prototype.constructor is correct.");

shouldBeEqualToString("Worklet.prototype.constructor.name", "Worklet");
</script>
</body>
</html>
